optimus_project_mcp/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.6
PyYAML>=6.0.1
mcp>=1.13.0
httpx>=0.27
